+++ internal/databases/postgres/postgres.go
@@ -284,13 +284,13 @@ func InitDB(connection_string string) error {
             data JSONB
         );
         CREATE TABLE IF NOT EXISTS github.pulls (
-            id INT NOT NULL,
+            id BIGINT NOT NULL,
             repo VARCHAR(255) NOT NULL,
             data JSON,
             PRIMARY KEY (id, repo)
         );
         CREATE TABLE IF NOT EXISTS github.pullsTest (
-            id INT NOT NULL,
+            id BIGINT NOT NULL,
             repo VARCHAR(255) NOT NULL,
             data JSON,
             PRIMARY KEY (id, repo)
